chore: move everthing to app component
